Reverses dependency for Portal-Core for Tests
- Portal is dependent of Core, and that is now the case for tests.
