+++ Digipost.Signature.Api.Client.Core.Tests/Digipost.Signature.Api.Client.Core.Tests.csproj
@@ -110,10 +110,6 @@
       <Project>{2D13F732-6A85-4165-AE2F-A87C09750398}</Project>
       <Name>Digipost.Signature.Api.Client.Core</Name>
     </ProjectReference>
-    <ProjectReference Include="..\Digipost.Signature.Api.Client.Portal\Digipost.Signature.Api.Client.Portal.csproj">
-      <Project>{d03be351-ca1f-4aac-968a-d3a0d8cd5197}</Project>
-      <Name>Digipost.Signature.Api.Client.Portal</Name>
-    </ProjectReference>
     <ProjectReference Include="..\Digipost.Signature.Api.Client.Scripts\Digipost.Signature.Api.Client.Scripts.csproj">
       <Project>{24d1d03d-4f19-488e-8f3f-30dc6c5b7bee}</Project>
       <Name>Digipost.Signature.Api.Client.Scripts</Name>
